backend/requirements.txt
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Session==0.8.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
python-dotenv==1.0.1
requests==2.32.3
spotipy==2.24.0
urllib3==2.2.3
Werkzeug==3.1.3
ytmusicapi==1.11.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.151.0
